requirements.txt
Django==1.9.5
django-filter==0.13.0
djangorestframework==3.3.3
mysqlclient=1.3.7
